--- a/tree.go
+++ b/tree.go
@@ -192,19 +192,19 @@ func (n *node) addRoute(path string, handlers HandlersChain) {
 						if len(n.path) >= len(path) || path[len(n.path)] == '/' {
 							continue walk
 						}
-					} else {
-						// Wildcard conflict
-						pathSeg := path
-						if n.nType != catchAll {
-							pathSeg = strings.SplitN(path, "/", 2)[0]
-						}
-						prefix := fullPath[:strings.Index(fullPath, pathSeg)] + n.path
-						panic("'" + pathSeg +
-							"' in new path '" + fullPath +
-							"' conflicts with existing wildcard '" + n.path +
-							"' in existing prefix '" + prefix +
-							"'")
 					}
+					
+					// Wildcard conflict
+					pathSeg := path
+					if n.nType != catchAll {
+						pathSeg = strings.SplitN(path, "/", 2)[0]
+					}
+					prefix := fullPath[:strings.Index(fullPath, pathSeg)] + n.path
+					panic("'" + pathSeg +
+						"' in new path '" + fullPath +
+						"' conflicts with existing wildcard '" + n.path +
+						"' in existing prefix '" + prefix +
+						"'")
 				}
 
 				c := path[0]