I recently read httprouter code, and find httprouter updated some codes, I want gin should also update `tree.go`.
But this updated whether impact gin performance, I need to test later on.

please @javierprovecho @appleboy @austinheap review the merged from httprouter pull request, thanks!

- [x] performance test compare
- [x] add and update unit test case