@appleboy 2018-06-23
We should update the https://github.com/gin-gonic/go-http-routing-benchmark repo after the pr has been merged.