@thinkerou 2018-06-23
OK, thanks @appleboy first I post performance data compare old gin version, and update it.