@thinkerou 2018-06-23
I use [web-framework-benchmark](https://github.com/vishr/web-framework-benchmark) compare old version and new version, new version performance a little better.

I will use https://github.com/gin-gonic/go-http-routing-benchmark again compare.

```sh
➜  web-framework-benchmark git:(master) ✗ go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/vishr/web-framework-benchmark
BenchmarkGinNewStatic-8      	   30000	     52666 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinNewGitHubAPI-8   	   20000	     74617 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinNewGplusAPI-8    	  300000	      4263 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinNewParseAPI-8    	  200000	      8094 ns/op	    1421 B/op	      26 allocs/op
BenchmarkGinOldStatic-8      	   30000	     55186 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinOldGitHubAPI-8   	   20000	     76590 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinOldGplusAPI-8    	  300000	      4399 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinOldParseAPI-8    	  200000	      8340 ns/op	    1421 B/op	      26 allocs/op
PASS
ok  	github.com/vishr/web-framework-benchmark	15.086s

➜  web-framework-benchmark git:(master) ✗ go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/vishr/web-framework-benchmark
BenchmarkGinNewStatic-8      	   30000	     54921 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinNewGitHubAPI-8   	   20000	     73823 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinNewGplusAPI-8    	  300000	      4223 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinNewParseAPI-8    	  200000	      8192 ns/op	    1421 B/op	      26 allocs/op
BenchmarkGinOldStatic-8      	   30000	     55740 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinOldGitHubAPI-8   	   20000	     76673 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinOldGplusAPI-8    	  300000	      4561 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinOldParseAPI-8    	  200000	      8509 ns/op	    1421 B/op	      26 allocs/op
PASS
ok  	github.com/vishr/web-framework-benchmark	15.269s

➜  web-framework-benchmark git:(master) ✗ go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/vishr/web-framework-benchmark
BenchmarkGinNewStatic-8      	   30000	     53096 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinNewGitHubAPI-8   	   20000	     73949 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinNewGplusAPI-8    	  300000	      4080 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinNewParseAPI-8    	  200000	      8169 ns/op	    1421 B/op	      26 allocs/op
BenchmarkGinOldStatic-8      	   30000	     54356 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinOldGitHubAPI-8   	   20000	     77451 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinOldGplusAPI-8    	  300000	      4539 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinOldParseAPI-8    	  200000	      8211 ns/op	    1421 B/op	      26 allocs/op
PASS
ok  	github.com/vishr/web-framework-benchmark	15.057s

➜  web-framework-benchmark git:(master) ✗ go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/vishr/web-framework-benchmark
BenchmarkGinNewStatic-8      	   30000	     55189 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinNewGitHubAPI-8   	   20000	     71915 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinNewGplusAPI-8    	  300000	      4250 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinNewParseAPI-8    	  200000	      8191 ns/op	    1421 B/op	      26 allocs/op
BenchmarkGinOldStatic-8      	   30000	     56249 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinOldGitHubAPI-8   	   20000	     76769 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinOldGplusAPI-8    	  300000	      4457 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinOldParseAPI-8    	  200000	      8163 ns/op	    1421 B/op	      26 allocs/op
PASS
ok  	github.com/vishr/web-framework-benchmark	15.204s

➜  web-framework-benchmark git:(master) ✗ go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/vishr/web-framework-benchmark
BenchmarkGinNewStatic-8      	   30000	     52571 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinNewGitHubAPI-8   	   20000	     73440 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinNewGplusAPI-8    	  300000	      4133 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinNewParseAPI-8    	  200000	      7873 ns/op	    1421 B/op	      26 allocs/op
BenchmarkGinOldStatic-8      	   30000	     56260 ns/op	    8693 B/op	     157 allocs/op
BenchmarkGinOldGitHubAPI-8   	   20000	     77832 ns/op	   10616 B/op	     203 allocs/op
BenchmarkGinOldGplusAPI-8    	  300000	      4254 ns/op	     681 B/op	      13 allocs/op
BenchmarkGinOldParseAPI-8    	  200000	      8197 ns/op	    1421 B/op	      26 allocs/op
PASS
ok  	github.com/vishr/web-framework-benchmark	14.918s

➜  web-framework-benchmark git:(master) ✗
```