@thinkerou 2018-06-23
I use https://github.com/julienschmidt/go-http-routing-benchmark, they are no obvious change, so I think the performance test pass.

- old version branch:

```sh
➜  go-http-routing-benchmark git:(master) go test -bench="Gin"
#GithubAPI Routes: 203
   Gin: 52672 Bytes

#GPlusAPI Routes: 13
   Gin: 3968 Bytes

#ParseAPI Routes: 26
   Gin: 6912 Bytes

#Static Routes: 157
   Gin: 30448 Bytes

goos: darwin
goarch: amd64
pkg: github.com/julienschmidt/go-http-routing-benchmark
BenchmarkGin_Param        	20000000	        65.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	20000000	       112 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	 5000000	       277 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	10000000	       134 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	20000000	        83.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	10000000	       131 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	   50000	     26113 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	20000000	        70.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	20000000	        83.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	20000000	       106 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 1000000	      1197 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	20000000	        67.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	20000000	        72.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	20000000	        85.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1000000	      2041 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	  100000	     18319 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/julienschmidt/go-http-routing-benchmark	28.299s
```

- new version branch:

```sh
➜  go-http-routing-benchmark git:(master) go test -bench="Gin"
#GithubAPI Routes: 203
   Gin: 52672 Bytes

#GPlusAPI Routes: 13
   Gin: 3968 Bytes

#ParseAPI Routes: 26
   Gin: 6912 Bytes

#Static Routes: 157
   Gin: 30448 Bytes

goos: darwin
goarch: amd64
pkg: github.com/julienschmidt/go-http-routing-benchmark
BenchmarkGin_Param        	20000000	        65.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	10000000	       111 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	 5000000	       270 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	10000000	       135 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	20000000	        82.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	10000000	       133 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	   50000	     27157 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	20000000	        73.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	20000000	        82.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	20000000	       105 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 1000000	      1235 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	20000000	        66.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	20000000	        71.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	20000000	        86.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1000000	      2152 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	  100000	     18602 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/julienschmidt/go-http-routing-benchmark	27.344s
```